# generated/openapi/cloudrouter-app-openapi.json
{
  "components": {
    "schemas": {
...
      "ProblemDetail": {
        "additionalProperties": true,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "i18nKey": {
            "description": "Optional stable localization key such as errors.result.40001.",
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "locale": {
            "description": "Optional effective BCP 47 locale used by framework message mapping.",
            "type": "string"
          },
          "status": {
            "maximum": 599,
            "minimum": 100,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "traceId": {
            "description": "Server-owned request correlation id.",
            "format": "uuid",
            "type": "string"
          },
          "type": {
            "format": "uri-reference",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "type": "object"
      },
...
      "SdkWorkApiResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "default": 0,
            "enum": [
              0
            ],
            "format": "int32",
            "maximum": 0,
            "minimum": 0,
            "type": "integer"
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "description": "Server-owned request correlation id.",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "type": "object"
      },
...
      "SdkWorkPlatformErrorCode": {
        "description": "Platform or domain error code per API_SPEC.md §15.3.",
        "format": "int32",
        "maximum": 79999,
        "minimum": 40001,
        "type": "integer"
      },
...
      "SettingsDataResponse": {
        "additionalProperties": false,
        "description": "Settings data response schema exposed by Cloud Router.",
        "name": "SettingsDataResponse",
        "properties": {
          "language": {
            "description": "Language field on settings data response.",
            "maxLength": 32,
            "type": "string"
          },
          "notifications": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SettingsNotifications"
              }
            ],
            "description": "Notifications field on settings data response."
          },
          "timezone": {
            "description": "Timezone field on settings data response.",
            "maxLength": 64,
            "type": "string"
          },
          "webhookUrl": {
            "description": "Webhook url field on settings data response.",
            "maxLength": 1024,
            "type": "string"
          }
        },
        "required": [
          "language",
          "timezone",
          "webhookUrl",
          "notifications"
        ],
        "type": "object"
      },
...
      "SiteRuntimeRetrieveResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "additionalProperties": false,
            "description": "Object schema used by site runtime retrieve result.",
            "properties": {
              "data": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SiteRuntimeSettingsResponse"
                  }
                ],
                "description": "Data field on site runtime retrieve result."
              }
            },
            "required": [
              "data"
            ],
            "type": "object"
          }
        ],
        "description": "Site runtime retrieve result schema exposed by Cloud Router.",
        "x-operation-id": "site.runtime.retrieve"
      },
...
      "UpdateSettingsRequest": {
        "additionalProperties": false,
        "description": "Update settings request schema exposed by Cloud Router.",
        "name": "UpdateSettingsRequest",
        "properties": {
          "language": {
            "description": "Language field on update settings request.",
            "maxLength": 32,
            "type": "string"
          },
          "notifications": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UpdateSettingsNotificationsRequest"
              }
            ],
            "description": "Notifications field on update settings request."
          },
          "timezone": {
            "description": "Timezone field on update settings request.",
            "maxLength": 64,
            "type": "string"
          },
          "webhookUrl": {
            "description": "Webhook url field on update settings request.",
            "maxLength": 1024,
            "type": "string"
          }
        },
        "required": [
          "language",
          "timezone",
          "webhookUrl",
          "notifications"
        ],
        "type": "object"
      },
...
      "UsersSettingsRetrieveResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "additionalProperties": false,
            "description": "Object schema used by users settings retrieve result.",
            "properties": {
              "data": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SettingsDataResponse"
                  }
                ],
                "description": "Data field on users settings retrieve result."
              }
            },
            "required": [
              "data"
            ],
            "type": "object"
          }
        ],
        "description": "Users settings retrieve result schema exposed by Cloud Router.",
        "x-operation-id": "users.settings.retrieve"
      },
      "UsersSettingsUpdateResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "additionalProperties": false,
            "description": "Object schema used by users settings update result.",
            "properties": {
              "data": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateSettingsResponse"
                  }
                ],
                "description": "Data field on users settings update result."
              }
            },
            "required": [
              "data"
            ],
            "type": "object"
          }
        ],
        "description": "Users settings update result schema exposed by Cloud Router.",
        "x-operation-id": "users.settings.update"
      }
    },
    "securitySchemes": {
      "AccessToken": {
        "in": "header",
        "name": "Access-Token",
        "type": "apiKey"
      },
      "AuthToken": {
        "bearerFormat": "SDKWork auth token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Generated from generated/api/api-contract-manifest.json for SdkworkAppClient.",
    "title": "SDKWork Cloud Router App API",
    "version": "0.1.0"
  },
...
  "openapi": "3.1.2",
  "paths": {
...
    "/app/v3/api/iam/users/settings": {
      "get": {
        "description": "List settings. Reads iam_user_preference, integration_webhook_endpoint, ops_notification_delivery. Writes none. File targets none.",
        "operationId": "users.settings.retrieve",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersSettingsRetrieveResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Server Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Error response."
          }
        },
        "security": [
          {
            "AccessToken": [],
            "AuthToken": []
          }
        ],
        "summary": "List settings",
        "tags": [
          "iam"
        ],
        "x-contract-kind": "read",
        "x-file-targets": [],
        "x-read-sources": [
          "iam_user_preference",
          "integration_webhook_endpoint",
          "ops_notification_delivery"
        ],
        "x-route-scope": "console",
        "x-sdk-domain": "iam",
        "x-sdkwork-api-authority": "sdkwork-cloudrouter-app-api",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-domain": "iam",
        "x-sdkwork-owner": "sdkwork-cloudrouter",
        "x-sdkwork-resource": "users.settings",
        "x-source-file": "apps/sdkwork-cloudrouter-pc/packages/sdkwork-cloudrouter-pc-console-settings/src/settingsService.ts",
        "x-write-tables": []
      },
      "put": {
        "description": "Update settings. Reads iam_user_preference, integration_webhook_endpoint. Writes iam_user_preference, integration_webhook_endpoint. File targets none.",
        "operationId": "users.settings.update",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          },
          "description": "Typed request payload for update settings.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersSettingsUpdateResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Server Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Error response."
          }
        },
        "security": [
          {
            "AccessToken": [],
            "AuthToken": []
          }
        ],
        "summary": "Update settings",
        "tags": [
          "iam"
        ],
        "x-contract-kind": "update",
        "x-file-targets": [],
        "x-read-sources": [
          "iam_user_preference",
          "integration_webhook_endpoint"
        ],
        "x-route-scope": "console",
        "x-sdk-domain": "iam",
        "x-sdkwork-api-authority": "sdkwork-cloudrouter-app-api",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-domain": "iam",
        "x-sdkwork-owner": "sdkwork-cloudrouter",
        "x-sdkwork-resource": "users.settings",
        "x-source-file": "apps/sdkwork-cloudrouter-pc/packages/sdkwork-cloudrouter-pc-console-settings/src/settingsService.ts",
        "x-write-tables": [
          "iam_user_preference",
          "integration_webhook_endpoint"
        ]
      }
    },
...
    "/app/v3/api/system/site/runtime": {
      "get": {
        "description": "List site branding. Reads ops_audit_log. Writes none. File targets none.",
        "operationId": "site.runtime.retrieve",
        "parameters": [
          {
            "description": "Tenant code query parameter.",
            "in": "query",
            "name": "tenant_code",
            "required": false,
            "schema": {
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "description": "Organization code query parameter.",
            "in": "query",
            "name": "organization_code",
            "required": false,
            "schema": {
              "maxLength": 64,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteRuntimeRetrieveResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Server Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Error response."
          }
        },
        "security": [],
        "summary": "List site branding",
        "tags": [
          "system"
        ],
        "x-contract-kind": "read",
        "x-file-targets": [],
        "x-read-sources": [
          "ops_audit_log"
        ],
        "x-route-scope": "public",
        "x-sdk-domain": "system",
        "x-sdkwork-api-authority": "sdkwork-cloudrouter-app-api",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-auth-mode": "anonymous",
        "x-sdkwork-domain": "system",
        "x-sdkwork-owner": "sdkwork-cloudrouter",
        "x-sdkwork-resource": "site.runtime",
        "x-source-file": "apps/sdkwork-cloudrouter-pc/packages/sdkwork-cloudroutes-pc-commons/src/siteBranding.ts",
        "x-write-tables": []
      }
    }
  },
  "security": [
    {
      "AccessToken": [],
      "AuthToken": []
    }
  ],
  "servers": [
    {
      "description": "Local app API server",
      "url": "http://localhost:18082"
    }
  ],
...
}